requirements.txt
google-auth == 2.38.0
google-auth-oauthlib == 1.2.1
google-auth-httplib2 == 0.2.0
google-api-python-client == 2.166.0
google.auth == 2.38.0
streamlit == 1.44.1
google-generativeai == 0.8.4
dotenv == 0.9.9
groq == 0.25.0